Move away from Paulo F. Oliveira and gha-example
We're assuming ErlEF will take this up, in the future
We're not making any assumptions on behaviour yet
(minor tweaks to the README add some TBD, though)
