+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -15,7 +15,7 @@ A clear and concise description of what the bug is.
 
 A list of software versions where the bug is apparent, as detailed as possible:
 
-* `gha-example`: ...
+* `setup-elp`: ...
 * other (where applicable): ...
 
 #### How to replicate
