+++ .github/pull_request_template.md
@@ -5,4 +5,4 @@ A brief description of your changes.
 Closes #&lt;issue&gt;.
 
 - [ ] I have performed a self-review of my changes
-- [ ] I have read and understood the [contributing guidelines](/paulo-ferraz-oliveira/gha-example/blob/main/CONTRIBUTING.md)
+- [ ] I have read and understood the [contributing guidelines](/paulo-ferraz-oliveira/setup-elp/blob/main/CONTRIBUTING.md)
