+++ README.md
@@ -1,41 +1,30 @@
-# `gha-example` [![Action][ci-img]][ci]
+# `setup-elp` [![Action][ci-img]][ci]
 
-[ci]: https://github.com/paulo-ferraz-oliveira/gha-example/actions/workflows/ci.yml
-[ci-img]: https://github.com/paulo-ferraz-oliveira/gha-example/actions/workflows/ci.yml/badge.svg
+[ci]: https://github.com/paulo-ferraz-oliveira/setup-elp/actions/workflows/ci.yml
+[ci-img]: https://github.com/paulo-ferraz-oliveira/setup-elp/actions/workflows/ci.yml/badge.svg
 
 ## The action
 
-This repository contains a Node.js-based GitHub Actions' template, with:
-
-* CI action files - `.github/workflows`
-* issues (bug report / feature request) - `.github/ISSUE_TEMPLATE`
-* pull request description - `.github/pull_request_template`
-* `src` - template source with input/output manipulation
-* `test` - tests
-* lint/format elements:
-  * `.markdownlint.yml`
-  * `.prettierrc`
-  * `.yamllint.yml`
-  * `eslint.config.js`
-
-It also contains project-support files as mentioned in section The project, in this README.
-
-To use it simply clone the project and search for references:
-
-* `paulo.ferraz.oliveira` (and `@gmail.com`, and or `Paulo`)
-* `gha-example`
-
-and tweak those to your liking.
-
-Note: for more JavaScript action -related information be sure to check out GitHub's documentation on
-[Creating a JavaScript action](https://docs.github.com/en/actions/creating-actions/creating-a-javascript-action).
+`setup-elp` allows you to type check your Erlang code, in your CI pipelines, using
+[eqWAlizer](https://github.com/WhatsApp/eqwalizer), with
+[Erlang Language Platform](https://github.com/WhatsApp/erlang-language-platform).
 
 ### Usage
 
 See [action.yml](action.yml) for the action's specification.
 
-Check out [GitHub's documentation](https://docs.github.com/en/actions/creating-actions/metadata-syntax-for-github-actions)
-to learn more about this file and its format.
+In your CI pipelines, use it considering CLI `elp` is available as an executable, as specified
+in the Erlang Language Platform repository mentioned above.
+
+#### Inputs
+
+The action supports the following inputs:
+
+TBD.
+
+#### Architectures
+
+TBD.
 
 ## The project
 
@@ -45,14 +34,14 @@ A complete changelog can be found under [GitHub releases](https://github.com/pau
 
 ### Code of Conduct
 
-This project's code of conduct is made explicit in [CODE_OF_CONDUCT.md](https://github.com/paulo-ferraz-oliveira/gha-example/blob/main/CODE_OF_CONDUCT.md).
+This project's code of conduct is made explicit in [CODE_OF_CONDUCT.md](https://github.com/paulo-ferraz-oliveira/setup-elp/blob/main/CODE_OF_CONDUCT.md).
 
 ### Contributing
 
 First of all, thank you for contributing with your time and patience.
 
 If you want to request a new feature make sure to
-[open an issue](https://github.com/paulo-ferraz-oliveira/gha-example/issues) so we can
+[open an issue](https://github.com/paulo-ferraz-oliveira/setup-elp/issues) so we can
 discuss it first.
 
 Bug reports and questions are also welcome, but do check you're using the latest version of the
@@ -62,12 +51,12 @@ might have already been answered before.
 Contributions will be subject to the MIT License.
 You will retain the copyright.
 
-For more information check out [CONTRIBUTING.md](https://github.com/paulo-ferraz-oliveira/gha-example/blob/main/CONTRIBUTING.md).
+For more information check out [CONTRIBUTING.md](https://github.com/paulo-ferraz-oliveira/setup-elp/blob/main/CONTRIBUTING.md).
 
 ### License
 
-License information can be found inside [LICENSE](https://github.com/paulo-ferraz-oliveira/gha-example/blob/main/LICENSE.md).
+License information can be found inside [LICENSE](https://github.com/paulo-ferraz-oliveira/setup-elp/blob/main/LICENSE.md).
 
 ### Security
 
-This project's security policy is made explicit in [SECURITY.md](https://github.com/paulo-ferraz-oliveira/gha-example/blob/main/SECURITY.md).
+This project's security policy is made explicit in [SECURITY.md](https://github.com/paulo-ferraz-oliveira/setup-elp/blob/main/SECURITY.md).
