+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "gha-example",
+  "name": "setup-elp",
   "license": "MIT",
   "private": true,
   "scripts": {
@@ -7,11 +7,11 @@
     "yamllint": "yamllint .github/workflows/**.yml && yamllint .*.yml && yamllint *.yml",
     "eslint": "eslint src/**/*.js && eslint test/**/*.js",
     "clean-dist": "rm -rf ./dist",
-    "build": "ncc build src/gha-example.js --no-cache --license licenses.txt",
+    "build": "ncc build src/setup-elp.js --no-cache --license licenses.txt",
     "format": "prettier src/**/*.js --write && prettier test/**/*.js --write && prettier *.js --write",
     "lint": "npm run markdownlint && npm run yamllint && npm run eslint",
     "build-dist": "npm run clean-dist && npm install && npm run build && npm run format && npm run lint",
-    "test": "node test/gha-example.test.js"
+    "test": "node test/setup-elp.test.js"
   },
   "dependencies": {
     "@actions/core": "1.10.1"
